# use type inference world in `return_type`