I found that one large source of latency in code loading is that we're repeatedly compiling the type inference code when it's called from `return_type`, since the type inference code in the system image is for a different world. This instead re-uses the precompiled world.

Significantly helps #15048. With this, precompiling DataFrames goes from 77 to 50 seconds, and precompiling JuliaDB goes from 130 to 75 seconds.