@ararslan 2018-12-20
Backport eligible for 1.0.x or 1.1?